refactor(vulnsource): extract the vulnerability source seam
Dependency scanning has always asked one question over the network — "is
this package version known to be vulnerable?" — and has only ever had one
possible answerer, wired in place. queryOSV was a package-level function
with unexported wire types, called inline from ScanArtifacts, with the
analyzer holding OSVBaseURL and an *http.Client. Nothing was swappable.

Introduce core/vulnsource: Query, Record, and a batch-shaped Source
interface. core/vulnsource/osv implements it against OSV.dev.

The refactor rule is that everything which is an OSV *protocol* artifact
moves behind the interface, and only the semantic contract stays visible:

  - ecosystem filtering and index remapping (OSV 400s a whole batch on one
    unknown ecosystem, which once zeroed out every Go/npm/PyPI result in a
    repo that merely had a Dockerfile)
  - batching at 1000 queries per request
  - the query-then-hydrate two-step, since /v1/querybatch returns only
    {id, modified} and severity mapping depends on what it omits
  - in-place slice hydration, because rebuilding from a flattened slice
    would attribute vulnerabilities to the wrong packages

What stays in deps is the layer applied to whatever a source returns:
severity mapping, fix selection, reachability scoping, upgrade commands.
A source cannot acquire different treatment by being a different source.

Source is batch-shaped rather than the per-package Lookup(ecosystem, pkg,
version) sketched in docs/design/intelligence-service.md. That signature
reads better and discards the batching real vulnerability APIs are built
around, turning one round trip into one per dependency.

Record is OSV-wire-compatible, json tags included, and the deps-side
record types are aliases rather than conversions. OSV's schema is the
interchange baseline every source must speak; a conversion layer here
would be a no-op that could only drift. The OSV batch types are exported
for the same reason — a server implementing the protocol shares these
definitions instead of restating them.

Behaviour is unchanged, and the guarantee is checked rather than assumed:
scanning a corpus with 114 VULN-001 findings before and after produces
identical finding sets (same sha256 over sorted fingerprints).

Degradation semantics are preserved exactly. An unreachable endpoint
still records degrade.OSV and returns partial results, because a silent
empty result is indistinguishable from "no vulnerabilities found".
WithOSVDisabled, WithOSVBaseURL, and OSVConfig are untouched.

The wire tests move to the package that now owns the wire protocol. New
tests cover the seam itself: that WithSource is actually consulted, that
its records get the same severity treatment as OSV's, that a source which
cannot answer fails the scan rather than returning a clean result, and
that the default source still reads OSVBaseURL at scan time.
